Fix to the bug that caused the menu to not be able to expand in mobile view.
The bug was that when the user was on a mobile device, bootstrap
automatically changes the 'myNavBar' div to a collapsable menu, however,
when the user clicked the menu button to expand the menu, nothing
happened.

The reason nothing happened was because there wasn't any JavaScript code
for it to work with, so therefore when you clicked the button, it
expected to run some JavaScript code but there wasn't any for it to run.
Therefore by adding the latest JQuery CDN, which Bootstrap requires, it
fixes the problem. Voila!